# docs/openapi.json
{
  "components": {
...
    "schemas": {
...
      "ErrorModel": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/ErrorModel.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "code": {
            "description": "Stable machine error code",
            "type": "string"
          },
          "detail": {
            "description": "A human-readable explanation specific to this occurrence of the problem.",
            "examples": [
              "Property foo is required but is missing."
            ],
            "type": "string"
          },
          "errors": {
            "description": "Optional list of individual error details",
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "field_errors": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object"
          },
          "instance": {
            "description": "A URI reference that identifies the specific occurrence of the problem.",
            "examples": [
              "https://example.com/error-log/abc123"
            ],
            "format": "uri",
            "type": "string"
          },
          "message": {
            "description": "Safe user message",
            "type": "string"
          },
          "request_id": {
            "type": "string"
          },
          "status": {
            "description": "HTTP status code",
            "examples": [
              400
            ],
            "format": "int64",
            "type": "integer"
          },
          "title": {
            "description": "A short, human-readable summary of the problem type. This value should not change between occurrences of the error.",
            "examples": [
              "Bad Request"
            ],
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "description": "A URI reference to human-readable documentation for the error.",
            "examples": [
              "https://example.com/errors/example"
            ],
            "format": "uri",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "InviteRequest": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/InviteRequest.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "display_name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "org_id": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "station_id": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "display_name",
          "role",
          "station_id",
          "org_id"
        ],
        "type": "object"
      },
...
      "PasswordResetResult": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/PasswordResetResult.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "link": {
            "type": "string"
          }
        },
        "required": [
          "link"
        ],
        "type": "object"
      },
...
      "RoleHistoryEvent": {
        "additionalProperties": false,
        "properties": {
          "actor": {
            "type": "string"
          },
          "after": {
            "type": "boolean"
          },
          "before": {
            "type": "boolean"
          },
          "created_at": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "sequence": {
            "format": "int64",
            "type": "integer"
          },
          "station": {
            "type": "string"
          },
          "station_id": {
            "type": "string"
          },
          "target": {
            "type": "string"
          }
        },
        "required": [
          "sequence",
          "actor",
          "target",
          "station_id",
          "station",
          "role",
          "before",
          "after",
          "created_at"
        ],
        "type": "object"
      },
...
      "UpdateUserRequest": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/UpdateUserRequest.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "display_name": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          }
        },
        "required": [
          "display_name",
          "enabled"
        ],
        "type": "object"
      },
      "UserView": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/UserView.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "display_name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "roles": {
            "items": {
              "$ref": "#/components/schemas/RoleView"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "stations": {
            "items": {
              "$ref": "#/components/schemas/StationView"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "user_id": {
            "type": "string"
          },
          "username": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "user_id",
          "email",
          "username",
          "display_name",
          "enabled",
          "roles",
          "stations"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      },
      "sessionCookie": {
        "description": "httpOnly session cookie",
        "in": "cookie",
        "name": "pomkita_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "PomKita API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/v1/users": {
      "get": {
        "operationId": "listUsers",
        "parameters": [
          {
            "explode": false,
            "in": "query",
            "name": "org_id",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UserView"
                  },
                  "type": [
                    "array",
                    "null"
                  ]
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "example": {
                    "summary": "Error example",
                    "value": {
                      "code": "INVALID_TOKEN",
                      "detail": "Request failed",
                      "field_errors": null,
                      "request_id": "request-id",
                      "status": 400,
                      "title": "Bad Request",
                      "type": "about:blank"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "List users",
        "x-permitted-roles": [
          "Owner",
          "Superadmin"
        ],
        "x-stable-error-codes": {
          "401": [
            "invalid_session"
          ],
          "403": [
            "user_administration_forbidden",
            "organization_scope_forbidden"
          ],
          "422": [
            "invalid_user_administration_request"
          ]
        }
      },
      "post": {
        "operationId": "inviteUser",
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "example": {
                  "summary": "Request example",
                  "value": {
                    "display_name": "Operator",
                    "email": "operator@example.test",
                    "role": "Operator",
                    "station_id": "22222222-2222-4222-8222-222222222222"
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/InviteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "example": {
                    "summary": "Error example",
                    "value": {
                      "code": "INVALID_TOKEN",
                      "detail": "Request failed",
                      "field_errors": null,
                      "request_id": "request-id",
                      "status": 400,
                      "title": "Bad Request",
                      "type": "about:blank"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Forbidden"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "Invite a user",
        "x-permitted-roles": [
          "Owner",
          "Superadmin"
        ],
        "x-stable-error-codes": {
          "403": [
            "user_administration_forbidden",
            "organization_scope_forbidden"
          ],
          "409": [
            "email_conflict",
            "username_conflict"
          ],
          "422": [
            "organization_has_no_station",
            "invalid_identity_request"
          ]
        }
      }
    },
    "/api/v1/users/{id}": {
      "get": {
        "operationId": "getUser",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserView"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "example": {
                    "summary": "Error example",
                    "value": {
                      "code": "INVALID_TOKEN",
                      "detail": "Request failed",
                      "field_errors": null,
                      "request_id": "request-id",
                      "status": 400,
                      "title": "Bad Request",
                      "type": "about:blank"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "Read a user",
        "x-permitted-roles": [
          "Owner",
          "Superadmin"
        ],
        "x-stable-error-codes": {
          "401": [
            "invalid_session"
          ],
          "403": [
            "user_administration_forbidden",
            "protected_user_forbidden",
            "csrf_required"
          ],
          "404": [
            "user_not_found"
          ],
          "422": [
            "invalid_user_administration_request"
          ]
        }
      },
      "patch": {
        "operationId": "updateUser",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "example": {
                  "summary": "Request example",
                  "value": {
                    "display_name": "Updated user",
                    "enabled": true
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserView"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "example": {
                    "summary": "Error example",
                    "value": {
                      "code": "INVALID_TOKEN",
                      "detail": "Request failed",
                      "field_errors": null,
                      "request_id": "request-id",
                      "status": 400,
                      "title": "Bad Request",
                      "type": "about:blank"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "Update a user",
        "x-csrf-required": true,
        "x-permitted-roles": [
          "Owner",
          "Superadmin"
        ],
        "x-stable-error-codes": {
          "401": [
            "invalid_session"
          ],
          "403": [
            "user_administration_forbidden",
            "protected_user_forbidden",
            "csrf_required"
          ],
          "404": [
            "user_not_found"
          ],
          "422": [
            "invalid_user_administration_request"
          ]
        }
      }
    },
    "/api/v1/users/{id}/password-reset": {
      "post": {
        "operationId": "resetUserPassword",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasswordResetResult"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "example": {
                    "summary": "Error example",
                    "value": {
                      "code": "INVALID_TOKEN",
                      "detail": "Request failed",
                      "field_errors": null,
                      "request_id": "request-id",
                      "status": 400,
                      "title": "Bad Request",
                      "type": "about:blank"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "Issue a user password reset",
        "x-csrf-required": true,
        "x-permitted-roles": [
          "Owner",
          "Superadmin"
        ],
        "x-stable-error-codes": {
          "401": [
            "invalid_session"
          ],
          "403": [
            "user_administration_forbidden",
            "protected_user_forbidden",
            "csrf_required"
          ],
          "404": [
            "user_not_found"
          ],
          "422": [
            "disabled_user_forbidden",
            "invalid_user_administration_request"
          ]
        }
      }
    },
    "/api/v1/users/{id}/role-history": {
      "get": {
        "operationId": "userRoleHistory",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RoleHistoryEvent"
                  },
                  "type": [
                    "array",
                    "null"
                  ]
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "example": {
                    "summary": "Error example",
                    "value": {
                      "code": "INVALID_TOKEN",
                      "detail": "Request failed",
                      "field_errors": null,
                      "request_id": "request-id",
                      "status": 400,
                      "title": "Bad Request",
                      "type": "about:blank"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "Read user role history",
        "x-permitted-roles": [
          "Owner",
          "Superadmin"
        ],
        "x-stable-error-codes": {
          "401": [
            "invalid_session"
          ],
          "403": [
            "user_administration_forbidden"
          ],
          "404": [
            "user_not_found"
          ]
        }
      }
    },
...
  },
...
}